requirements.txt
imageio==2.17.0
matplotlib==3.5.1
mmcv-full==1.5.0
numpy==1.21.2
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.0.1
prettytable==3.3.0
PyYAML==6.0
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.0
tifffile==2022.4.8
timm==0.5.4
torch==1.10.0
torchvision==0.11.1
tqdm==4.64.0
wandb==0.12.14
